Now relies on arduino-cli
1.
Updates am into am2 and Makefile contents, so that arduino-builder is no
longer used to compile and upload in the command line.
From now on, it'll be done with arduino-cli.

2.
Also fixes a minor bug where a declaration was missing inside
RF433recv.h. Strangely, arduino-builder would not produce any error, but
arduino-cli does (arduino-cli is right though: the declaration was
missing indeed).